move haversine module to sphere
