+++ clouddrift/__init__.py
@@ -5,5 +5,4 @@ __version__ = version
 from clouddrift.raggedarray import RaggedArray
 import clouddrift.adapters
 import clouddrift.analysis
-import clouddrift.haversine
 import clouddrift.sphere
