+++ docs/api.rst
@@ -33,13 +33,6 @@ Analysis
   :members:
   :undoc-members:
 
-Haversine
----------
-
-.. automodule:: clouddrift.haversine
-  :members:
-  :undoc-members:
-
 RaggedArray
 -----------
